requirements.txt
attrs==21.4.0
importlib-metadata==4.4
pluggy>=1.0.0
py>=1.11.0
pytest>=6.2.5
allure-pytest>=2.9.45
pytest-rerunfailures>=4.1
pytest-testrail-e2e==2.2.3
pytest-testconfig>=0.1.0
pytest-logger>=0.4.0
pytest-xdist>=2.5.0
pytest-order>=1.0.1
selenium>=4.3.0
zipp==3.5.0
Pillow>=8.4.0
requests>=2.27.1
pytest-timeout>=2.1.0
mysql-connector-python>=8.0.15
boto3>=1.24.11
botocore>=1.27.11
opencv-python>=4.6.0
numpy~=1.23.1
matplotlib~=3.5.2
